Fixed complex coordinate system scaling and added complex value to rendered arrow translation

+++ visualizeFourierSeries/src/visualizeFourierSeries/arrowAndCircleRenderData.java
@@ -9,10 +9,10 @@ public class arrowAndCircleRenderData {
 	private int y_pixel_coordinate;
 	private Polygon arrow_head_polygon;
 	private Point body_arrow_connection;
-	private int arrow_body_stroke;
+	private float arrow_body_stroke;
 	private int circle_radius;
 	
-	public arrowAndCircleRenderData(int x, int y, Polygon arrow_head_poly, Point body_arrow_con, int arrow_body_str, int circ_radius){
+	public arrowAndCircleRenderData(int x, int y, Polygon arrow_head_poly, Point body_arrow_con, float arrow_body_str, int circ_radius){
 		x_pixel_coordinate = x;
 		y_pixel_coordinate = y;
 		arrow_head_polygon = arrow_head_poly;
@@ -37,7 +37,7 @@ public class arrowAndCircleRenderData {
 		return body_arrow_connection;
 	}
 	
-	public int getArrowBodyStroke(){
+	public float getArrowBodyStroke(){
 		return arrow_body_stroke;
 	}
